+++ src/utils/sort.js
@@ -1,3 +1,5 @@
+import { isoStrToDate, utcStrToDate } from './formater'
+
 const getStateWeight = (state) => {
   switch (state) {
     case 'healthy':
@@ -28,28 +30,15 @@ export const getPropValue = (obj, prop) => {
   }
   return obj[prop]
 }
-const isoStrToDate = (isoStr) => {
-  const reg = /^(\d{4}-\d{2}-\d{2})T(\d{2}:\d{2}:\d{2})Z$/
-  const results = isoStr.match(reg)
-  if (results && results.length === 3) {
-    const d = results[1].split('-').map(item => parseInt(item, 10))
-    const t = results[2].split(':').map(item => parseInt(item, 10))
-    return new Date(Date.UTC(d[0], d[1] - 1, d[2], t[0], t[1], t[2]))
-  }
-  return 'Invalid ISO Date'
-}
+
 export function sortVolume(dataSource) {
   dataSource.sort((a, b) => {
     const status = getStateWeight(a.state) - getStateWeight(b.state)
     if (status === 0) {
-      return new Date(b.created).getTime() - new Date(a.created).getTime()
+      return utcStrToDate(b.created).getTime() - utcStrToDate(a.created).getTime()
     }
     return status
   })
-}
-
-export function sortVolumeByName(dataSource) {
-  dataSource.sort((a, b) => a.name.localeCompare(b.name))
 }
 
 export function sortVolumeBackups(dataSource) {
@@ -73,6 +62,18 @@ export function sortTable(a, b, prop) {
 export function sortTableByISODate(a, b, prop) {
   const valueA = isoStrToDate(getPropValue(a, prop))
   const valueB = isoStrToDate(getPropValue(b, prop))
+  if (valueA < valueB) {
+    return -1
+  }
+  if (valueA > valueB) {
+    return 1
+  }
+  return 0
+}
+
+export function sortTableByUTCDate(a, b, prop) {
+  const valueA = utcStrToDate(getPropValue(a, prop))
+  const valueB = utcStrToDate(getPropValue(b, prop))
   if (valueA < valueB) {
     return -1
   }
